Usunięcie starszych wersji plików: main, reset_hasla, logowanie
