+++ baza_ksiazek.txt
@@ -29,4 +29,4 @@
 29;Virginia Woolf;Do latarni morskiej;14.0
 30;Marcel Proust;W poszukiwaniu straconego czasu;3.0
 31;Joseph Conrad;Jądro ciemności;6.0
-32;test;test;0.0
+
