+++ baza_wyrejestrowani.txt
@@ -1,2 +1,3 @@
 002;a;a;100;a@;dupa;a;sadf
 004;GREG;GREGory;55;jacek@o2.pl;admin;admin22;sport
+024;zo;z;20;z@.;z;z;dramat;True
